Added /xm fav command, this will display off of a players favorites.
Added /xm fav <radioid> command, this toggles a favorite id, it also
checks if the id is valid.

Added /xm #, this command will attempt to auto join a radio, normal
verification checks are performed so the player has to be able to join
with out a password before this function works, additionally
range/capacity checks are performed as well.